streamlit/requirements.txt
streamlit[snowflake]>=1.54.0
snowflake-connector-python>=3.3.0
requests>=2.31.0
